Fix an OOB write in sparse minmax selectivity estimation (#8933)
Made an off-by-one error. It appeared only on old chunks where the max
column is the last column in the compressed chunk tables. Such chunks
can't be created on modern versions because the metadata columns now go
before the corresponding data column. This is why it was not caught by
any tests.