Bump macos to 15
GitHub is cancelling the job on MacOS 13 with the following messages:

This is a scheduled macos-13 brownout. The macOS-13 based runner images
are being deprecated. For more details, see
https://github.com/actions/runner-images/issues/13046.

The macOS-13 based runner images are being deprecated, consider
switching to macOS-15 (macos-15-intel) or macOS 15 arm64 (macos-latest)
instead. For more details see
https://github.com/actions/runner-images/issues/13046