Remove Continuous Aggregate old format support
In release 2.7.0 (2022-05-24) we introduced a new CAgg format without
storing partials. So starting in the version was possible to create
either the old and new format by using the `finalized` option when
creating a new CAgg.

In release 2.8.0 (2022-08-31) we added the `cagg_migrate` procedure for
users migrate their old format CAggs into the new format.

In release 2.12.0 (2023-09-26) we restricted creating CAggs in the old
format starting on PG15.

More than three years passed since we introduced the new format that
optimized significantly query performance and storage leading customers
to cost savings. So is time to say good bye to the old format and
completely remove support for it from our code base. It means if for
some reason a user still have the old format the extension update will
be blocket, so is good to use the `cagg_migrate` procedure to migrate to
the old format and make sure to really remove the old format CAgg and
then try update the extension again.

Related PRs: #4294 #4269 #4941

Related Releases:
https://github.com/timescale/timescaledb/releases/tag/2.7.0
https://github.com/timescale/timescaledb/releases/tag/2.8.0
https://github.com/timescale/timescaledb/releases/tag/2.12.0