Bump bloom filter metadata column type to 2 (#8761)
The bloom1 sparse indexes hashing depends on build options. This leads
to index corruption when people mix different builds. We observed this
with some customers in cloud.

Hashing is also different between our current arm and amd64 packages.

This PR makes the bloom filter index column name dependent on the
hashing scheme, therefore avoiding the incompatibility. The old prefix
is "bloom1", the new options are "bloomH" (modern hashing we also use
for vectorized aggregation) and "bloomG" (standard postgres hashing).

After this change, the old "bloom1" sparse index column will be ignored,
so the bloom filter indexes will stop working for the chunks created
before the migration. The index can be restored by decompressing the
chunk and compressing it again.