Add GUC for default chunk interval
When auto-tuning of the chunk interval is available, we'd like to be
able to set a small default interval and tune into a bigger one (going
from big to small is not great because big intervals lock you in). When
auto-tuning is not available, it makes sense that the default is bigger.

With this change, a default chunk interval can be set with a new GUC
timescaledb.default_chunk_interval. The GUC expects a text-format
PostgreSQL INTERVAL type (e.g., '1 day' or '1 week'). It is only
possible to set the default chunk interval for hypertables that use an
INTERVAL-compatible partitioning type, i.e., TIMESTAMP(TZ), DATE, and
UUID (version 7). The default value of the GUC is "null", which means
the current behavior is preserved (i.e., the hard-coded defaults are
used).