Ensure no XID is assigned during chunk query
When querying hypertables, the dimension slices of the queried chunks
are locked using a row-level (tuple) lock. This will automatically
assign a transaction ID to the query because PostgreSQL stores row-level
locks in the tuple (i.e., on disk) instead of in memory. Writing to the
tuple requires a transaction ID and a lock on the transaction.

However, these additional locks could create problems since they might
block other operations. For example, it prevents creating replication
slots as long as the locks are held.

This change fixes the tuple lock issue for cases when chunks are queried
directly so that replication slots can be created while the chunk is
read.