Prevent `drop_chunks` from dropping CAgg-dependent chunks
This PR changes `drop_chunks` behaviour to skip dropping chunks that
contain data required for continuous aggregate refreshes by checking
the CAgg watermark. A notice is emitted for each skipped chunk.

This is mainly to prevent data retention policies (which use
drop_chunks under-the-hood) from dropping chunks which haven't been
refreshed yet by a corresponding refresh policy.

This also introduces a 'force' option to `drop_chunks` to allow users to
override the watermark-protection and drop chunks containing unrefreshed
data. Using 'force' to drop a chunk containing unrefreshed CAgg data
generates a warning for the user.

Note that invalidation logs are not checked. It is still possible to
drop chunks containing unrefreshed data that was backfilled.