Use baserestrictinfo for chunk exclusion (#8987)
Not sure why we're going through the parse tree here again.

Our custom code probably went out of sync with PG starting with version 16, so chunk exclusion is not working in some cases. It's compensated for by the standard Postgres chunk exclusion, so there is no visible effect, but we shouldn't really be doing that because it leads to planning performance regressions.