New default value for buckets per batch
In the Continuous Aggregate refresh policy we have an option
`buckets_per_batch=1` intented to split the refresh window in small
batches based on the bucket size and process the refresh in multiple
transactions to don't spike resource consumption.

In the past we decided to be conservative use the default value of `1`
to don't risk having batches processing too much data, but it
demostrated be inefficient increasing too much the total execution
time of the refresh policy.

Isn't easy to determine a good default value here because it is totally
dependent of the dataset distribution, but we want to keep this feature
on by default so we're increasing it to `10` in order to have more data
to process in each batch.

We're working to have a smart way to define this size (maybe becoming
adaptative) but for now the value keep static.