Improve approximate row count on compressed chunks
In #8654 we fixed the `approximate_row_count` for compressed chunks to
use Postgres statistics doing the math `pg_class.reltuples * 1000`.

This is not very accurate since we can have less than 1000 rows per
segment leading to bad information, so exposed the internal compressed
batch size estimation in an SQL function to be used to properly
calculate the approximate row count.