Improve parallel worker calculation for compressed chunks
Postgres calculates the number of parallel workers a query requires
based on the number of heap pages of the relations its querying. Since
compressed chunks store most of their data is toasted away, heap pages
tend to be on the low side so we need to account for the compression
factor of the heap. This is not perfect but it brings us closer to
the estimated parallel worker numbers of an uncompressed chunk.
This change only affects parallel sequential scans since they will benefit
most from increased parallelism.