Prepare for expressions in vectorized grouping and aggregation (#9029)
In the vectorized aggregation executor node, we used to use the input
column offsets to reference the grouping and aggregation columns.
Prepare to use the full expressions there by storing a full Expr* node.
The evaluation of the expression is changed to happen before the actual
call of the grouping or aggregate function, inside the grouping policy,
to allow for better memory usage patterns.