Fix check for pending ranges in CAgg refresh
In #8607 we added an extra check for pending ranges after processing
invalidations and materialize the data for self healing in case of
interruped CAgg refresh left behind some pending ranges.

The problem is we're not using the latest snapshot to check for pending
ranges leading to unecessary executing the check since the deleted
pending range still visibile. Fixed it by pushing the latest snapshot
and then don't see the deleted pending range.