Fix relfrozenxid not updated after splitting chunk
When splitting a chunk in two, one new chunk relation is created and the
tuples are split across the new and old. The old relation had its
relfrozenxid updated in pg_class via a heap swap, but relfrozenxid
wasn't updated on the new relation and instead it was always set to the
most recent xid. However, sometimes relfrozenxid needs to be held back
due to concurrent transactions, but the relfrozenxid on the new chunk
was not updated to the "lower" relfrozenxid.

This could lead to an error during VACUUM:

"ERROR:  found xmin 831 from before relfrozenxid 832"

Fix this issue by properly setting relfrozenxid on the new chunk after
the split.

(cherry picked from commit 095f9d3fb17bfa707ad9f5a72791d5948ec3197f)