Add test to trigger missing chunk column stats bug
If the hypertable cache is populated with an entry while chunk skipping
is off, the range space information is omitted. This cached entry can
remain after chunk skipping is turned on, leading to a bug where chunk
column stats are not created for new chunks.

Add a test function to manually clear the hypertable cache and add a
test to trigger the bug using the function.

(cherry picked from commit ec3ea2764696a947aafc969ca11f78ca0171072a)