Only include ARM64 when custom runner is present
The custom GitHub runner for ARM64 is only present in the main
TimescaleDB GitHub repository, so when the workflow is run in another
repository the job hangs indefinitely while waiting for a runner to pick
it up.

To make it possible to test and debug the workflows in a fork
repository, only include the custom runner job in case the workflow runs
in the main TimescaleDB repository.