Can apply multikey sort for columnstore when one numeric key is pinned to a Const of different type (#9069)
Fixes #9066.

When we have a columnstore with multiple segmentby keys and we have a
query ordered on columnstore keys where one segmentby key is pinned to a
Const `(key = Const)`, we failed to to treat this key as Const segmentby
when both key and Const were numeric compatible but different types like
int4 and int32.

The issue is fixed by properly checking that OpExpr `(key = Const)`
really is an equality operator comparing compatible types.